+++ src/neptune/internal/backends/hosted_neptune_backend.py
@@ -585,6 +585,9 @@ class HostedNeptuneBackend(NeptuneBackend):
         errors = list()
         assign_operations = list()
 
+        has_hash_exclude_metadata = self._client_config.has_feature(OptionalFeatures.ARTIFACTS_HASH_EXCLUDE_METADATA)
+        has_exclude_directories = self._client_config.has_feature(OptionalFeatures.ARTIFACTS_EXCLUDE_DIRECTORY_FILES)
+
         for op in artifact_operations:
             try:
                 artifact_hash = self.get_artifact_attribute(container_id, container_type, op.path).hash
@@ -600,6 +603,8 @@ class HostedNeptuneBackend(NeptuneBackend):
                         parent_identifier=container_id,
                         entries=op.entries,
                         default_request_params=DEFAULT_REQUEST_KWARGS,
+                        exclude_directory_files=has_exclude_directories,
+                        exclude_metadata_from_hash=has_hash_exclude_metadata,
                     )
                 else:
                     assign_operation = track_to_existing_artifact(
@@ -610,6 +615,7 @@ class HostedNeptuneBackend(NeptuneBackend):
                         parent_identifier=container_id,
                         entries=op.entries,
                         default_request_params=DEFAULT_REQUEST_KWARGS,
+                        exclude_directory_files=has_exclude_directories,
                     )
 
                 if assign_operation: